requirements.txt
discord.py
feedparser
python-dotenv
sumy
bs4
requests

